+++ Day2a/Worker.cs
@@ -4,7 +4,7 @@ namespace AdventOfCode2023.Day2a;
 
 public class Worker : IWorker
 {
-    public int DoWork(string inputFile)
+    public long DoWork(string inputFile)
     {
         var games = new List<Game>();
         foreach (var line in File.ReadLines(inputFile))
